committing player interface and human player implementation
